# swagger.yaml
consumes:
    - application/json
produces:
    - application/json
schemes:
    - https
swagger: "2.0"
info:
    contact:
        email: test@trahan.dev
        name: Justin Trahan
    license:
        name: N/A
    termsOfService: there are no TOS at this moment, use at your own risk we take no responsibility
    title: go-infra API.
    version: v1.1.0
host: infra.trahan.dev
paths:
...
    /host-server-types:
        get:
            operationId: GetAllHostServerTypes
            responses:
                "200":
                    $ref: '#/responses/GetAllHostServerTypesResponse'
                "401":
                    description: Unauthorized
                "500":
                    description: Internal Server Error
            summary: Get all available host server types.
            tags:
                - host-servers
    /host-servers:
        get:
            operationId: GetAllHostServers
            responses:
                "200":
                    $ref: '#/responses/HostServersResponse'
                "401":
                    description: Unauthorized
                "500":
                    description: Internal Server Error
            summary: Get all host servers.
            tags:
                - host-servers
    /host-servers/create:
        post:
            operationId: CreateHostServer
            parameters:
                - in: body
                  name: body
                  schema:
                    $ref: '#/definitions/CreateHostServerRequest'
                  x-go-name: Body
            responses:
                "200":
                    $ref: '#/responses/HostServerResponse'
                "400":
                    description: Invalid request
                "401":
                    description: Unauthorized
                "500":
                    description: Internal Server Error
            summary: Create a new host server.
            tags:
                - host-servers
    /host-servers/{ID}:
        delete:
            operationId: DeleteHostServer
            parameters:
                - format: uuid
                  in: path
                  name: ID
                  required: true
                  type: string
            responses:
                "200":
                    description: Host server deleted successfully
                "400":
                    description: Invalid request
                "401":
                    description: Unauthorized
                "404":
                    description: Not Found
                "500":
                    description: Internal Server Error
            summary: Delete a host server.
            tags:
                - host-servers
        get:
            operationId: GetHostServer
            responses:
                "200":
                    $ref: '#/responses/HostServerResponse'
                "400":
                    description: Invalid request
                "401":
                    description: Unauthorized
                "404":
                    description: Not Found
                "500":
                    description: Internal Server Error
            summary: Get a host server by ID.
            tags:
                - host-servers
        put:
            operationId: UpdateHostServer
            parameters:
                - format: uuid
                  in: path
                  name: ID
                  required: true
                  type: string
                - in: body
                  name: body
                  schema:
                    $ref: '#/definitions/UpdateHostServerRequest'
                  x-go-name: Body
            responses:
                "200":
                    $ref: '#/responses/HostServerResponse'
                "400":
                    description: Invalid request
                "401":
                    description: Unauthorized
                "404":
                    description: Not Found
                "500":
                    description: Internal Server Error
            summary: Update a host server.
            tags:
                - host-servers
...
    /platform-types:
        get:
            operationId: GetAllPlatformTypes
            responses:
                "200":
                    $ref: '#/responses/GetAllPlatformTypesResponse'
                "401":
                    description: Unauthorized
                "500":
                    description: Internal Server Error
            summary: Get all available platform types.
            tags:
                - host-servers
...
definitions:
...
    CreateHostServerRequest:
        properties:
            host_server_type_ids:
                description: Host server type IDs that this server supports
                example:
                    - 123e4567-e89b-12d3-a456-426614174000
                    - 123e4567-e89b-12d3-a456-426614174001
                items:
                    format: uuid
                    type: string
                type: array
                x-go-name: HostServerTypeIDs
            hostname:
                description: Hostname of the server
                example: server-01.example.com
                type: string
                x-go-name: Hostname
            ip_address:
                description: IP address of the server
                example: 192.168.1.100
                type: string
                x-go-name: IPAddress
            platform_type_ids:
                description: Platform type IDs that this server supports
                example:
                    - 123e4567-e89b-12d3-a456-426614174002
                    - 123e4567-e89b-12d3-a456-426614174003
                items:
                    format: uuid
                    type: string
                type: array
                x-go-name: PlatformTypeIDs
            ssh_key_id:
                description: SSH key ID for authentication
                example: 123e4567-e89b-12d3-a456-426614174000
                format: uuid
                type: string
                x-go-name: SSHKeyID
            sudo_password_token_id:
                description: Optional sudo password token ID
                example: 123e4567-e89b-12d3-a456-426614174001
                format: uuid
                type: string
                x-go-name: SudoPasswordTokenID
            username:
                description: Username for SSH connection
                example: admin
                type: string
                x-go-name: Username
        required:
            - hostname
            - ip_address
        type: object
        x-go-package: github.com/babbage88/go-infra/services/host_servers
...
    HostServerResponse:
        properties:
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            host_server_types:
                items:
                    $ref: '#/definitions/HostServerType'
                type: array
                x-go-name: HostServerTypes
            hostname:
                type: string
                x-go-name: Hostname
            id:
                format: uuid
                type: string
                x-go-name: ID
            ip_address:
                type: string
                x-go-name: IPAddress
            last_modified:
                format: date-time
                type: string
                x-go-name: LastModified
            platform_types:
                items:
                    $ref: '#/definitions/PlatformType'
                type: array
                x-go-name: PlatformTypes
            ssh_key_id:
                format: uuid
                type: string
                x-go-name: SSHKeyID
            sudo_password_token_id:
                format: uuid
                type: string
                x-go-name: SudoPasswordTokenID
            username:
                type: string
                x-go-name: Username
        title: HostServerResponse represents a host server response.
        type: object
        x-go-package: github.com/babbage88/go-infra/services/host_servers
    HostServerType:
        properties:
            id:
                description: Unique identifier for the host server type
                example: 123e4567-e89b-12d3-a456-426614174000
                format: uuid
                type: string
                x-go-name: ID
            last_modified:
                description: Last modification timestamp
                example: "2024-01-15T10:30:00Z"
                format: date-time
                type: string
                x-go-name: LastModified
            name:
                description: Name of the host server type
                example: Database Server
                type: string
                x-go-name: Name
        required:
            - id
            - name
            - last_modified
        type: object
        x-go-package: github.com/babbage88/go-infra/services/host_servers
...
    PlatformType:
        properties:
            id:
                description: Unique identifier for the platform type
                example: 123e4567-e89b-12d3-a456-426614174001
                format: uuid
                type: string
                x-go-name: ID
            last_modified:
                description: Last modification timestamp
                example: "2024-01-15T10:30:00Z"
                format: date-time
                type: string
                x-go-name: LastModified
            name:
                description: Name of the platform type
                example: Docker Host
                type: string
                x-go-name: Name
        required:
            - id
            - name
            - last_modified
        type: object
        x-go-package: github.com/babbage88/go-infra/services/host_servers
...
    UpdateHostServerRequest:
        properties:
            host_server_type_ids:
                description: Host server type IDs that this server supports
                example:
                    - 123e4567-e89b-12d3-a456-426614174000
                    - 123e4567-e89b-12d3-a456-426614174001
                items:
                    format: uuid
                    type: string
                type: array
                x-go-name: HostServerTypeIDs
            hostname:
                description: Hostname of the server
                example: server-01.example.com
                type: string
                x-go-name: Hostname
            ip_address:
                description: IP address of the server
                example: 192.168.1.100
                type: string
                x-go-name: IPAddress
            platform_type_ids:
                description: Platform type IDs that this server supports
                example:
                    - 123e4567-e89b-12d3-a456-426614174002
                    - 123e4567-e89b-12d3-a456-426614174003
                items:
                    format: uuid
                    type: string
                type: array
                x-go-name: PlatformTypeIDs
            ssh_key_id:
                description: SSH key ID for authentication
                example: 123e4567-e89b-12d3-a456-426614174000
                format: uuid
                type: string
                x-go-name: SSHKeyID
            sudo_password_token_id:
                description: Optional sudo password token ID
                example: 123e4567-e89b-12d3-a456-426614174001
                format: uuid
                type: string
                x-go-name: SudoPasswordTokenID
            username:
                description: Username for SSH connection
                example: admin
                type: string
                x-go-name: Username
        type: object
        x-go-package: github.com/babbage88/go-infra/services/host_servers
...
responses:
...
    GetAllHostServerTypesResponse:
        description: (empty)
        schema:
            items:
                $ref: '#/definitions/HostServerType'
            type: array
    GetAllPlatformTypesResponse:
        description: (empty)
        schema:
            items:
                $ref: '#/definitions/PlatformType'
            type: array
...
    HostServerResponse:
        description: (empty)
        schema:
            $ref: '#/definitions/HostServerResponse'
    HostServersResponse:
        description: (empty)
        schema:
            items:
                $ref: '#/definitions/HostServerResponse'
            type: array
...
securityDefinitions:
    bearer:
        in: header
        name: Authorization
        type: apiKey
security:
    - bearer: []